Move `execInLoginShell` to utils
